fix: silence PHP 8.5 deprecation warnings and SSL CA attribute - Added explicit `deprecations` log channel - Fixed compatibility with PHP 8.5+ for SSL CA attribute - Replaced implicit-nullable params with explicit ones - Replaced non-canonical `(double)` and `(integer)` casts with `(float)` and `(int)`

+++ app/Http/Resources/AnimeResourceBasic.php
@@ -101,7 +101,7 @@ class AnimeResourceBasic extends JsonResource
         // Return the array
         return [
             'library' => [
-                'rating' => (double) $givenRating?->rating,
+                'rating' => (float) $givenRating?->rating,
                 'review' => $givenRating?->description,
                 'status' => $library?->status,
                 'rewatchCount' => $library?->rewatch_count,
